Rewrite commands system
